+++ packages/css-tokenizer/test/css-tokenizer/function-token.test.ts
@@ -1,7 +1,6 @@
 import { TokenType } from '../../src/common/enums/token-types';
 import { tokenize } from '../../src/css-tokenizer';
 import { getFormattedTokenName } from '../../src/utils/token-names';
-import { type TokenData } from '../helpers/test-interfaces';
 
 describe('function-token', () => {
     test.each([
@@ -19,10 +18,11 @@ describe('function-token', () => {
         String.raw`\\66 unc(`, // Extra whitespace
         String.raw`\\0066 unc(`, // Extra leading zeros and extra whitespace
     ])(`should tokenize '%s' as ${getFormattedTokenName(TokenType.Function)}`, (actual) => {
-        const tokens: TokenData[] = [];
-        tokenize(actual, (...args) => tokens.push(args));
-        expect(tokens).toEqual([
-            [TokenType.Function, 0, actual.length],
-        ]);
+        const onToken: jest.MockedFunction<OnTokenCallback> = jest.fn();
+
+        tokenize(actual, onToken);
+
+        expect(onToken).toHaveBeenCalledTimes(1);
+        expect(onToken.mock.calls[0]?.slice(0, 3)).toEqual([TokenType.Function, 0, actual.length]);
     });
 });