+++ packages/css-tokenizer/test/extended-css-tokenizer/extended-css-tokenizer.test.ts
@@ -10,7 +10,7 @@ describe('tokenizer', () => {
     test('should skip leading BOM in the extended tokenizer', () => {
         const tokens: TokenData[] = [];
         tokenizeExtended('\uFEFF ', (...args) => tokens.push(args));
-        expect(tokens).toEqual([
+        expect(tokens.map(([type, start, end]) => [type, start, end])).toEqual([
             [TokenType.Whitespace, 1, 2],
         ]);
     });
@@ -22,10 +22,11 @@ describe('tokenizer', () => {
                 const start = context.offset;
                 // this is a simple handler that just consumes one single code point - just for testing purposes
                 context.consumeCodePoint();
-                context.onToken(TokenType.Delim, start, context.offset);
+                context.onToken(TokenType.Delim, start, context.offset, undefined, context.stop);
             }],
         ]));
-        expect(tokens).toEqual([
+
+        expect(tokens.map(([type, start, end]) => [type, start, end])).toEqual([
             [TokenType.Colon, 0, 1],
             [TokenType.Function, 1, 8],
             [TokenType.Delim, 8, 9],