+++ packages/agtree/CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [3.0.0-alpha.4] - 2024-02-03
+
+### Changed
+
+- Updated [@adguard/css-tokenizer] to `v1.2.0` which introduces the `hasToken` function.
+
+[3.0.0-alpha.4]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-3.0.0-alpha.4
+
 ## [3.0.0-alpha.3] - 2025-01-30
 
 ### Changed